Added multiple dimensions
-Swap dimensions with /dim command
-Everything should work cross dimensionally?
-Fixed a bug where invalid socket on windows would overflow size_t
-Got claude to update some command signatures for me lol
-Added transferDimension function pointer to the commands... A little jank but its needed
-Fixed some out of bounds wrapping for inventory serialization / loading
-Made a server shutdown flag that is atomic across threads to prevent race conditions (spent way too long wondering why nbt data was getting corrupted)
-Fixed a bug where item tags weren't being serialized with their names

+++ src/bpp_shared/networking/network_stream.cpp
@@ -43,9 +43,9 @@ void NetworkStream::flushWriteBufferBlocking() {
 
     size_t sent = 0;
     while (sent < writeBuffer.size()) {
-        size_t result = send(client_socket,
+        int result = send(client_socket,
             reinterpret_cast<const char*>(writeBuffer.data() + sent),
-            (writeBuffer.size() - sent), 0);
+            static_cast<int>(writeBuffer.size() - sent), 0);
         if (result <= 0) break;
         sent += static_cast<size_t>(result);
     }
@@ -221,10 +221,10 @@ bool NetworkStream::flushWriteBuffer() {
     if (writeBuffer.empty()) return connected;
     size_t sent = 0;
     while (sent < writeBuffer.size()) {
-        size_t result = send(
+        int result = send(
             client_socket,
             reinterpret_cast<const char*>(writeBuffer.data() + sent),
-            writeBuffer.size() - sent,
+            static_cast<int>(writeBuffer.size() - sent),
             0
         );
         if (result < 0) {
